Version 2.0
This is a complete update of the code for CSAD. The updated code is as simple as possible, and the intention is to ease the process for students to use CSAD for experiments. All previous code is removed, and the updated code is built in LabVIEW 2017, VeriStand 2017 and Matlab 2016b.